P3.50: dead_code on no_state_axis, float_cmp on the radius assert_ne!
Two more clippy errors from the same crowbar-app bin (non-test) build,
each masked by the other until the first was fixed.

no_state_axis: read only by the invariant test in surface.rs -- unlike
Surface::root (the precedent this follows), nothing in the driver/
snapshot path calls it either, so the correct guard is `not(test)`
alone, not root's `not(any(feature = "driver", test))`. Applied at
both concrete definition sites: SurfaceParams's default body and the
blanket `impl<T: SurfaceParams> Params for T`'s forwarding body. Not
touched: the abstract signature in the Params trait itself (no body,
nothing to be dead) and workspace_branch_icon.rs's own override
(apparently not flagged -- trait-impl overrides reachable through
dyn Params dispatch aren't the same lint target as a default body).

row_layout/repo_avatar.rs:75 (clippy::float_cmp): neither sibling's
case -- not two compile-time consts (cluster 2's .to_bits() case), and
not a value proved exactly 0.0 by construction (cluster 3's case).
record.radius and FULL_RADIUS are both laid-out `Pixels`, converted to
raw f32 before comparing for no reason the surrounding code needed --
line 74 two lines above already compares `Pixels` directly, and
avatar.rs's own test compares FULL_RADIUS against px(f32::MAX) the
same way. Dropped both f32::from() conversions; the typed comparison
doesn't trip the lint.

Not run here: clippy. Static review only, per "do not run the gate
suite" -- grepped both files and their P3.50 neighbors for the same
patterns (no_state_axis call sites, f32::from comparisons) to confirm
each fix's scope and that no third instance was missed.

+++ native/crates/crowbar-app/src/row_layout/repo_avatar.rs
@@ -72,10 +72,13 @@ fn the_letter_shape_is_a_rounded_square_not_a_circle(cx: &mut TestAppContext) {
     let theme = Theme::DARK;
     let record = find(&records, "repo-avatar");
     assert_eq!(record.radius, theme.radius_sm.value());
-    assert_ne!(
-        f32::from(record.radius),
-        f32::from(crowbar_ui::components::avatar::FULL_RADIUS),
-    );
+    // `Pixels`, not the raw `f32` either side would convert to — the same
+    // typed comparison line 74 above already makes, and the one `avatar.rs`'s
+    // own `assert_eq!(FULL_RADIUS, px(f32::MAX))` makes too. Comparing the
+    // laid-out `Pixels` value directly is what avoids `clippy::float_cmp`
+    // here; converting both sides to `f32` first, the way this line used to,
+    // is what triggered it.
+    assert_ne!(record.radius, crowbar_ui::components::avatar::FULL_RADIUS);
 }
 
 /// Every size resolves to its own authored box — `16 × 16`, `20 × 20`,
